Change the loop index in normalizestring() to size_t too, to avoid a warning on Windows.

+++ Python/codecs.c
@@ -82,7 +82,7 @@ int PyCodec_Register(PyObject *search_function)
 static
 PyObject *normalizestring(const char *string)
 {
-    register int i;
+    register size_t i;
     size_t len = strlen(string);
     char *p;
     PyObject *v;
